chore: Cleaned up prefixed text input component
