+++ app/helpers/view_helper.rb
@@ -17,10 +17,10 @@ module ViewHelper
     flash: 'Viral::FlashComponent',
     help_text: 'Viral::Form::HelpTextComponent',
     icon: 'Viral::IconComponent',
-    input_group: 'Viral::Form::InputGroupComponent',
     pageheader: 'Viral::PageHeaderComponent',
     prefixed_checkbox: 'Viral::Form::Prefixed::CheckboxComponent',
     prefixed_select: 'Viral::Form::Prefixed::SelectComponent',
+    prefixed_text_input: 'Viral::Form::Prefixed::TextInputComponent',
     pill: 'Viral::PillComponent',
     select: 'Viral::Form::SelectComponent',
     tabs: 'Viral::TabsComponent',
